pyproject.toml
[project]
name = "navigate"
version = "0.1.0"
description = "A synapse module for creating anonymous class section-specific channels"
authors = [{ name = "ok-nick", email = "nick.libraries@gmail.com" }]
requires-python = ">3.11.4"
dependencies = ["matrix-synapse==1.95.1"]

[tool.pyright]
typeCheckingMode = "basic"

[tool.ruff]
fixable = ["ALL"]
extend-select = ["I"]
